% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_distinct.R
\name{assert_distinct}
\alias{assert_distinct}
\title{An assert statement to determine whether there are multiple rows by grouping id}
\usage{
assert_distinct(
  data_df,
  group_by_vars = NULL,
  error_context = get_current_file(),
  data_path = get_data_path(),
  error_path = get_metadata_path("AMPLIO_ERROR_PATH")
)
}
\arguments{
\item{data_df}{A dataframe}

\item{group_by_vars}{A string or vector of strings of column name(s) to group by in asserting uniqueness}

\item{error_context}{For internal use}

\item{data_path}{For internal use}

\item{error_path}{For internal use}
}
\value{
Non-unique rows by grouping variables
}
\description{
An assert statement to determine whether there are multiple rows by grouping id
}
