% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_phi.R
\name{check_for_phi}
\alias{check_for_phi}
\title{check_for_phi}
\usage{
check_for_phi(consecutive_digits = 2, regex_pattern = NULL)
}
\arguments{
\item{consecutive_digits}{tbd}

\item{regex_pattern}{tbd}
}
\description{
Draft version of searching site for PHI
}
